R8-c 实跑 --stage=full 与 smoke.sh：修 8 个缺陷（其中 4 个让 compose 从来起不来）
工作包：R8-c（作者排通 Docker Hub 后补跑最后两项验证）

栈起不来的四个
1. bitnami/kafka:3.7 清单不存在——Bitnami 2025 年把旧镜像迁到 bitnamilegacy/ 归档仓库。
   教材要用几年，不宜钉在归档镜像上，改用官方 apache/kafka:3.7.2；
   环境变量去掉 bitnami 特有的 CFG_ 段，卷路径与健康检查同步调整。
   书中 8.6 的 compose 清单一并改，保持书与配套一致。
2. ./db 整个目录被挂进 /docker-entrypoint-initdb.d，Postgres 会按字母序执行其中
   所有 .sql，包括本该由宿主机 psql 手工跑的 load-s3-data.sql（读 ../datasets/*.csv）。
   容器里没有那些 CSV，初始化失败、容器 exit 3、整栈起不来。改为只挂 001、002。
3. 实体与建表类型不一致：reading.value 建表是 numeric（8.3 数据字典与 DDL 清单都如此），
   实体却是 Double，ddl-auto: validate 直接拒绝启动。改为 BigDecimal
   （也符合 5.5 节“带单位的小数应使用 BigDecimal”），ReadingEvent、ReadingDto 一并改。
4. smoke.sh 用 python3 解析登录响应，没有 python3 即以 127 退出，
   而 README 把它写成“一条命令验证”。改用 sed。

契约不符的三个（--stage=full 首跑 12/18）
5. 未认证返回 403 而非契约规定的 401。已配 authenticationEntryPoint 与
   accessDeniedHandler，401 与 403 分开并各自产出契约错误体。
   这条同时是 smoke.sh 第 2 步的断言。
6. GET …/readings 不校验 from < to，参数写反时返回 200 与空数组——
   调用方会把“参数写反”误读成“这段时间没有观测”。
   已按契约返回 400 与 {code:"INVALID_RANGE", field:"from"}。
7. 种子数据里没有“有对象但无观测”的测点，真实后端演示不出 204（只能演示 404），
   第4章的“暂无观测”状态在真机上无法复现。002_seed.sql 补 DAM-A-D-01（刻意不给观测）。

我自己引入的一个回归
8. R7-d 加的 ApiExceptionHandler 兜底 @ExceptionHandler(Exception.class) 把
   ResponseStatusException 也吞了，登录失败本该返回的 401 被翻译成 500。
   已补显式的 ResponseStatusException 处理器保留原状态码。

最终结果（全部实跑）
- --stage=full 18/18；smoke.sh 5 步全过；--stage=teaching 20/20；
  closeloop-check 20/20；--stage=lesson52 16/16；vitest 47/47；JUnit 5/5。
- 五个容器 postgres/redis/kafka/backend/frontend 均 healthy。

说明
- 验证时宿主机 8080 被 Docker 的陈旧端口预留占住（系统层无监听，也不在 Windows
  排除段内）。为不打断本机上其他项目的容器，未重启 Docker，改用临时 compose
  覆盖把前端发到 18080；该覆盖文件在临时目录，未进仓库，
  仓库里的 docker-compose.yml 仍是 8080。

门禁
- check_textbook.py --strict --build 通过；check_listings.py 9 组一致；
  mkdocs build --strict 零警告。
- 字数不变（235,923），本包只改代码与配置。

+++ companion/water-platform-demo/backend/src/main/java/edu/example/qingyuan/ReadingEntity.java
@@ -1,5 +1,6 @@
 package edu.example.qingyuan;
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 import jakarta.persistence.*;
 
@@ -10,13 +11,15 @@ public class ReadingEntity {
     // 幂等键：数据库层为 (occurred_at, event_id) 复合唯一索引
     //（TimescaleDB 超表的唯一索引必须包含分区列，见第8章 8.3 节）
     @Column(name = "event_id", nullable = false) private String eventId;
-    @Column(name = "value") private Double value;
+    // 对应 001_init.sql 的 numeric（见 8.3 数据字典）。用 Double 会被 ddl-auto: validate 拒绝，
+    // 也不符合 5.5 节“带单位的小数应使用 BigDecimal”的口径。
+    @Column(name = "value") private BigDecimal value;
     @Column(nullable = false) private String unit;
     @Column(nullable = false) private String quality;
     @Column(nullable = false) private String source;
 
     protected ReadingEntity() {}
-    public ReadingEntity(ReadingId id, String eventId, Double value,
+    public ReadingEntity(ReadingId id, String eventId, BigDecimal value,
                          String unit, String quality, String source) {
         this.id = id; this.eventId = eventId; this.value = value;
         this.unit = unit; this.quality = quality; this.source = source;
@@ -29,7 +32,7 @@ public class ReadingEntity {
 
     public ReadingId getId() { return id; }
     public String getEventId() { return eventId; }
-    public Double getValue() { return value; }
+    public BigDecimal getValue() { return value; }
     public String getUnit() { return unit; }
     public String getQuality() { return quality; }
     public String getSource() { return source; }
